flask/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Flash
Flask-Bcrypt
Flask-Login
Flask-Mail
Pillow
gunicorn
eventlet
flask-script
manage
coverage
flask_testing
bs4
requests
pandas
psycopg2-binary
greenlet
pyopenssl
